Rename DcConfig and create sub directory llm
